get data script

+++ extract_test_from_brown.sh
@@ -6,7 +6,10 @@ SECTIONS="01 02 03"
 
 mkdir -p $CONLL05/test.brown/synt
 
-rm $CONLL05/test.brown/synt/test.brown.synt.gz
+# Remove output file if it exists, since we're appending
+if [ -e $CONLL05/test.brown/synt/test.brown.synt.gz ]; then
+    rm $CONLL05/test.brown/synt/test.brown.synt.gz
+fi
 
 for section in $SECTIONS; do
     cat $BROWN/CK/CK$section.MRG | awk '{if($1 !~ "*x*") print}' | $SRLCONLL/bin/wsj-removetraces.pl | $SRLCONLL/bin/wsj-to-se.pl -w 0 -p 1 | gzip >> $CONLL05/test.brown/synt/test.brown.synt.gz
